Updated exception handling for circuit breaker is order-service. Implement circuit breaker pattern customer-service. Added a success check of the order deletion request for a customer. Fixed minor issues.

+++ customer-service/src/main/java/pytaichuk/customer_service/service/CustomerService.java
@@ -1,19 +1,26 @@
 package pytaichuk.customer_service.service;
 
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
+import io.github.resilience4j.retry.annotation.Retry;
 import lombok.RequiredArgsConstructor;
-import pytaichuk.customer_service.dto.customer.CustomerResponse;
-import pytaichuk.customer_service.dto.customer.CustomerRequest;
-import pytaichuk.customer_service.dto.customer.search.CustomerSearchRequest;
-import pytaichuk.customer_service.dto.customer.search.CustomerSearchResponse;
-import pytaichuk.customer_service.dto.order.OrderResponse;
-import pytaichuk.customer_service.model.Customer;
-import pytaichuk.customer_service.repository.CustomerRepository;
-import pytaichuk.customer_service.exception.FindException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
+import pytaichuk.customer_service.dto.customer.CustomerRequest;
+import pytaichuk.customer_service.dto.customer.CustomerResponse;
+import pytaichuk.customer_service.dto.customer.search.CustomerSearchResponse;
+import pytaichuk.customer_service.dto.order.OrderResponse;
+import pytaichuk.customer_service.exception.CustomerDeleteException;
+import pytaichuk.customer_service.exception.FindException;
+import pytaichuk.customer_service.model.Customer;
+import pytaichuk.customer_service.repository.CustomerRepository;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -21,6 +28,7 @@ import java.util.List;
 public class CustomerService {
     private final WebClient.Builder webClientBuilder;
     private final CustomerRepository customerRepository;
+    private final Logger log = LoggerFactory.getLogger(CustomerService.class);
 
     @Transactional
     public void createCustomer(CustomerRequest customerRequest){
@@ -38,21 +46,31 @@ public class CustomerService {
         customerRepository.save(customer);
     }
 
+    @Retry(name = "orderRetry")
+    @CircuitBreaker(name = "order")
     @Transactional
     public void delete(Long id){
         Customer customer = findCustomerById(id);
 
         System.out.println(customer.getId());
 
-        webClientBuilder.build().delete()
+        ResponseEntity<Void> responseEntity = webClientBuilder.build().delete()
                 .uri("http://order-service/api/v1/order/byCustomer/{customerId}", customer.getId())
                 .retrieve()
                 .toBodilessEntity()
                 .block();
 
-        customerRepository.delete(customer);
+        HttpStatusCode httpStatusCode = Objects.requireNonNull(responseEntity).getStatusCode();
+
+        if(httpStatusCode.isSameCodeAs(HttpStatusCode.valueOf(200))){
+            customerRepository.delete(customer);
+        } else {
+            throw new CustomerDeleteException("Помилка при видаленні клієнта. Спробуйте пізніше.;\n");
+        }
     }
 
+    @Retry(name = "orderRetry")
+    @CircuitBreaker(name = "order")
     public CustomerSearchResponse findCustomer(String telephoneNumber){
 
         Customer customer = findCustomerByTelephoneNumber(telephoneNumber);
